Fix NoneType int comparison errors for blueair i35 humidifier (#136)
Apparently we couldn't fetch the filter_usage yet thus the property is None.
But because it is apparently called by all of the actions (maybe unintionally?) all actions fail with this type error on i35. I tested with this fix fan on/off and led on off works with an i35.