Use coordinator as the name for *Coordinator.
Also removes the redundant self._device member because
super().__init__()
already set self.coordinator.

The name is consistent with the naming convention from the ha developer
docs:

https://developers.home-assistant.io/docs/integration_fetching_data/#coordinated-single-api-poll-for-data-for-all-entities

(also the Entity class examples appear to still use the confusing
'device' as the name of the updater; seems like coordinators were
introduced after that document was written).

For the next step, I am debating whether to expose the blueair_api.*Device as
coordinator.device, since adding the type annotation propagation is a
lot of work with no obvious gain in neither layer cleanness or
functionatliy.

But this change is large enough that I would like to file it first.