A few experiements for anitiflicker.
https://github.com/dahlb/ha_blueair/issues/164

Seems like only the change enabling the callbacks -> ha state update
fixed the flickering of entity states on the UI.

However we are amplifying writes to the HA db with the callback
triggering state updates for all of the entites using the
coordinator/device. We will most likely be required to fix this if we
are to be included in ha core (people seem to care about minimizing the
size of the db).